requirements.txt
django~=1.11
django-braces~=1.13.0